Database encryption works! \o/
Database is automagically created with regular websql when logging in as a new user. Password is stored in memory and database is automatically saved as .encrypted upon app shutdown.

delete the original database, start the app again, press login and the database is auto decrypted.

next up: delete unencrypted database on shutdown.